requirements.txt
# Core
numpy<2
scipy>=1.16.1
scikit-learn>=1.7.1
sentence-transformers>=5.1.0
faiss-cpu>=1.12.0       # Use faiss-gpu if you want GPU indexing
google-generativeai
google-cloud-storage

# File handling
PyPDF2>=3.0.1
python-dotenv>=1.1.1
requests>=2.32.5

# FastAPI (if you plan to serve API)
fastapi>=0.116.1
uvicorn>=0.35.0

# Optional
tqdm>=4.67.1
pydantic>=2.11.7